Standardize the minute and seconds component of Kafka generated traffic files.
After doing quite some research, it seems that Oozie does not support globs
in filename patterns. This makes it close to impossible to run an Oozie
coordinator job as there are no deterministic rules to determine the minute
and seconds component.

This small patch contains a simple function to create standardized timestamps.
It accepts a frequency parameter that indicates how often to collapse minutes.