Standardize the minute and seconds component of Kafka generated traffic files.
After doing quite some research, it seems that Oozie does not support globs
in filename patterns. This makes it close to impossible to run an Oozie
coordinator job as there are no deterministic rules to determine the minute
and seconds component.

This small patch just hardcodes the minute and second component to '00'. The
downside of doing this is that we ever decide to go store Kafka files at a
more granular timedelta then we would have to undo this patch.