Turn kraken-etl jar into shaded jar
We finally want to arrive at a shaded jar as it thereby suffices to
deploy a single etl jar into Hadoop / Kraken nodes.

So—although kraken-etl does not yet contain code and the resulting jar
is still empty—, we switch to a shaded jar already now, to allow
detecting problems (that are due to shading) early in the development
of kraken-etl.

Change-Id: I6ff3d85c9b622b9b89aa69346eada9b731c9e446